requirements.txt
# Requirements cho myPythonTool
# Cài đặt: pip install -r requirements.txt

# Tool compress-images
Pillow>=10.0.0

# Tool text-encoding-converter
chardet>=5.0.0

# Tool extract-archive (tùy chọn)
# Bỏ comment nếu cần hỗ trợ thêm định dạng nén
# py7zr>=0.20.0        # Hỗ trợ .7z
# rarfile>=4.0         # Hỗ trợ .rar (cần cài WinRAR/unrar)

